Sobha Highland Hosur Road Bangalore
Sobha Highland Bangalore is a newly launched project in the most developed city of India. Experience an urban oasis, this residential project embodies the spirit and lifestyle of Singapore- a city renowned for its joyous and vibrant culture. This premium complex has been designed by Sobha Limited.

Live in a project that has a 50% green area. Luxurious in excellent amenities, your futuristic home reflects the spirit of the island nation. This project will provide a life of complete enjoyment and luxury.

The developer is guaranteed to bring the best quality of life experience to the community in Bangalore. The project is located right next to the massive Sobha Highland Hosur Road apartment which is already occupied and thriving.

Sobha Highland floor plan offers 2BHK and 3BHK luxury residential apartments. This duplex apartment is beautifully spread over wide acres of land parcel with wonderful outdoor spaces. There are a whole lot of units in its development. This design of the property in Bangalore is keeping in mind the Vastu principles for correct face, measurement, space arrangement, layout, and spatial geometry.

Apartments in Development will offer you finely crafted apartments where the quality of finishes shines through in every development from start to finish. It is an ultimate reflection of the urban chic lifestyle in the heart of the city. May you be assured of excellent profits and growth in the coming future. This project will come with the most luxurious living between nature and city life.

The project is located at a prime location Hosur Road, Bengaluru, Karnataka. This new residence is located near the mall and is in close proximity to various communities and amenities. Hosur Road is well connected by metro and road making it an ideal location with good infra. Metro, cinema theatres, temples, restaurants, schools, and much more make the connectivity of the area even better.

For More Details
Visit us - https://bit.ly/3oNqDKr